renaming project from `descriptor` to `base-descriptor`
