+++ examples/slottable_data_descriptor_example.py
@@ -1,4 +1,4 @@
-from descriptor import (
+from base_descriptor import (
     BaseSlottableDataDescriptor,
     SlottableDefaultDescriptor,
 )
